Backbone of handling moves implemented and more!
Added new files:
-- Core Game:
+ Snake
- What will represent the Player, can be created as an AI or without.
Currently without doesn't support the player making moves.
+ MoveHandler
- The multiple-purpose thread that will Consume and Produce to the
BoundedBuffer within the game.
-- Testing utilities:
+ IllegalBoundedBufferException
+ IllegalRoleException

Repurposed the following files:
+ ClientGUI
- Is now what 'creates' an instance of MainThread that handles the Game
more closely.
+ MainThread
- Is now an extension of Thread, will be the runner of the Runnable
MoveHandler (Currently not working 100%).
+ Server
- Server has been changed to support two states: TESTING | PRODUCTION,
if in the Testing state it will be running test code, and in Production
we will run it like the game being ran.
Altered the following files:
+ Game
- Added some more debugging purposes (I have an idea on how to
intergrate Game with MoveHandler. Talk more on the 9th about it.
+ testConsumer, testProducer
- Removed TODO: