Added support for Multi-player humans
+ Added Files

+ util.MoveKey - An encapsulation of an Integer, this Integer is defined by
KeyEvent.VK_x where x is any Key you wish to assign to a snake.
For example, the default for player 1 is VK_UP, LEFT, DOWN, RIGHT, etc.

+ util.MoveSet - An encapsulation class of MoveKeys, a MoveSet contains 4
MoveKeys, see MoveKey.

~ Added support in the Following classes for the MoveSets:

~ Snake - Now updates it's own moves if it's a human player.

~ Game - KeyPressedEvent now polls the list of Human Players, if the key
is defined in their respective MoveSets, sent that snake the Key to
handle it.
Also, on game loop it now checks the list of humans to see if
they're dead, if so remove them from the list. (Possibly inefficient,
considering alternate placement).

x. Note:
1. ClientGUI is now expected to create a list of human players.

2. Game and Server now expect an ArrayList of humanPlayers on init.

3. MoveHandler no longer spawns BIG_FOOD, this is due to the Game no
longer keeping track of Score. Kieran stated he wishes to create a
Highscores list, which would remove the need of Score anyway, leaving
this design-choice up to him.

4. Refactored Snake's 'enemysnake' to be 'boardLocation', as that is
what it is used for.