Added support back for pausing.
> Modified following files

>Game

Added code for on KeyPress event as VK_SPACE, paused = !paused.

>MoveHandler

Added code for if Game.paused = true, delay for 10 ms (10 ms to fake the game automatically starting back up again).

Thoughts are that it's kinda a cheap way to pause, but the alternative of killing the threads then starting them up again doesn't work. Once a Thread exits, it goes into a 'dead' state that can only be removed via nulling it. Which is fine, no big deal. The issue is with starting the mainLoop() again. The mainLoop() contains a while(true) which means that the method that calls mainLoop is held up in mainLoop() this is a problem because our KeyEvent handler fires mainLoop(), therefore the players can never enter moves again, that's why we have a Server(). (Could change the Server to have pausing functionality, however this gets really awkward).