Added ability to differentiate Human Players via having a different coloured head.
Also added Logger support, files are saved to C:\tmp\test\@filename, the Logger uses a BoundedBuffer that contains LoggingPackets

+ Logger.java
This is the class that handles the BoundedBuffer<LoggingPacket>, a Logger outputs to a BufferedWriter, the current implementation is unable to clear the Buffer before the application closes, also doesn't handle heaps of players, using it at the moment just for clarifying concurrency bugs. (Which should mean we just use two Snakes to see the moves being handled.

+ LoggingPacket.java
A class that contains a String (A message) to be put to the Log file.

~ ClientGUI.java
Now utilizes the CopyOnWriteArrayList for true concurrency actions within the Server. This CopyOnWriteArrayList flows through the application, so that all instances of this ArrayList should be using the CopyOnWrite functionality.

Now also creates the Logger instance, which needs to be passed down the line for use Application Wide. (Note the constructor changes for the following classes
> Server.java
> Game.java
> MoveHandler.java

~ Game.java
Now supports the ability for us to quickly swap from pool threads to normal threads, also the option to use threaded AI snake creation.

~ MoveHandler.java
Now supports threaded snake creation. (Thinking of moving this functionality elsewhere).

~ Snake.java
Now supports a different "head", therefore we can change the "head" for the snake, which is a number. This number will represent the head in the grid.

Overall, message me if something doesn't make sense B^).